requirements.txt
streamlit==1.25.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.0.0
pytorch-grad-cam>=1.4.4
matplotlib>=3.4.0
seaborn>=0.11.0
gdown>=4.6.0
requests>=2.25.0
